ZOOKEEPER-1077: C client lib doesn't build on Solaris (Chris Nauroth via rgs)
git-svn-id: https://svn.apache.org/repos/asf/zookeeper/trunk@1679951 13f79535-47bb-0310-9956-ffa450edef68

+++ src/c/src/zk_log.c
@@ -155,8 +155,9 @@ void log_message(log_callback_fn callback, ZooLogLevel curLevel,
 
 #ifndef THREADED
 
+    // pid_t is long on Solaris
     ofs = snprintf(buf, FORMAT_LOG_BUF_SIZE,
-                   "%s:%d:%s@%s@%d: ", time, pid,
+                   "%s:%ld:%s@%s@%d: ", time, (long)pid,
                    dbgLevelStr[curLevel], funcName, line);
 #else
 
@@ -167,7 +168,7 @@ void log_message(log_callback_fn callback, ZooLogLevel curLevel,
     #endif
 
     ofs = snprintf(buf, FORMAT_LOG_BUF_SIZE-1,
-                   "%s:%d(0x%lx):%s@%s@%d: ", time, pid, tid,
+                   "%s:%ld(0x%lx):%s@%s@%d: ", time, (long)pid, tid,
                    dbgLevelStr[curLevel], funcName, line);
 #endif
